\name{getMeans}
\alias{getMeans}
\title{Get Mean Methylation}
\usage{
  getMeans(bmptab, eset)
}
\arguments{
  \item{bmptab}{The 'table' list item from the bumps object
  returned by \code{bumphunter}}

  \item{eset}{Usually a GenomicRatioSet, generated by
  minfi}
}
\value{
  A data.frame containing mean methylation data.
}
\description{
  Get mean methylation for a given region
}
\details{
  This is an internal function and not intended for direct
  use. The general idea is to get the mean expression of
  methylation probes in a region considered to be
  significantly differentially methylated, in order to
  create a dotplot.
}
\author{
  James W. MacDonald (\email{jmacdon@u.washington.edu})
}

