\name{getRegion}
\alias{getRegion}
\title{Get Methylation Region}
\usage{
  getRegion(bumpsObj, eset, row)
}
\arguments{
  \item{bumpsObj}{The 'table' list item from the bumps
  object returned by \code{bumphunter}}

  \item{eset}{Usually a GenomicRatioSet, generated by
  minfi}

  \item{row}{Which row of the table are we using?}
}
\value{
  A list with the max and min of the genomic region, as
  well as the Ensembl ID of any genes therein.
}
\description{
  Get methylation region
}
\details{
  This is an internal function and not intended to be
  called directly. The goal is to get a reasonably sized
  genomic region around an area that is differentially
  methylated, so the resulting plot looks good
}
\author{
  James W. MacDonald (\email{jmacdon@u.washington.edu})
}

